#VYE-273 #commnet moved paragraph compoenent to constants folder
