+++ src/applications/verify-your-enrollment/constants/index.js
@@ -1,4 +1,5 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import manifest from '../manifest.json';
 
 export const BASE_URL = `${manifest.rootUrl}/`;
@@ -35,6 +36,23 @@ export const blockURLsRegEx =
   '^((?!http|www\\.|\\.co|\\.net|\\.gov|\\.edu|\\.org).)*$';
 
 // export const STREET_LINE_MAX_LENGTH = 20;
+export const Paragraph = ({ title, date, className }) => {
+  return (
+    <p
+      className={`vads-u-font-size--md vads-u-font-family--serif vads-u-font-weight--bold ${className}`}
+    >
+      {title}:
+      <span className="vads-u-font-weight--normal vads-u-font-family--sans text-color vads-u-display--inline-block vads-u-margin-left--1">
+        {date}
+      </span>
+    </p>
+  );
+};
+Paragraph.propTypes = {
+  className: PropTypes.string,
+  date: PropTypes.string,
+  title: PropTypes.string,
+};
 
 export const ACTIVEDUTYBENEFITSSTATEMENT = (
   <p>
